re-arrange the con_k function and bug fix for docstrings
